:bug: Fixing compatibility issues
To list our projects, we must develop the following page: https://app.zeplin.io/project/5e21f3e014d75a576cbb16f9/screen/5e6e5304b9ae1b17d560de7a

In this commit, I've fixed some compatibility issues with the upstream repository.

Now, our events must be showing as expected.
