+++ components/EventsWithCategoryTitle/styles.js
@@ -1,0 +1,12 @@
+const styles = (theme) => ({
+  title: {
+    marginTop: theme.spaces.loose.value,
+    marginBottom: theme.spaces.loose.value,
+  },
+
+  list: {
+    paddingBottom: theme.spaces.loose.value,
+  },
+});
+
+export default styles;
