+++ containers/Events/index.js
@@ -6,7 +6,7 @@ import EventsGroupedByCategory from '../EventsGroupedByCategory';
 import { response } from './mockedAPI';
 
 const Events = () => (
-  <Layout topGutter={210}>
+  <Layout topGutter={150}>
     <EventsGroupedByCategory events={response} />
   </Layout>
 );
